requirements.txt
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
boto==2.49.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-rest-auth==2.1.4
Django==3.0.8
django-allauth==0.42.0
django-cors-headers==3.4.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-mysql==3.7.1
django-rest-auth==0.9.5
django-seed==0.2.2
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
Faker==4.1.1
fstring==1.7.4
google==3.0.0
google-api-core==1.21.0
google-auth==1.19.2
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-crc32c==0.1.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
swagger-spec-validator==2.7.3
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
xmltodict==0.12.0
